+++ Unity/UnityVideoProject/Assets/Plugins/iOS/NativeCallProxy.mm
@@ -26,13 +26,13 @@ public:
         newBufferDataAvailable_ = false;
     }
     
-    void getOutput(std::unique_ptr<uint32_t[]>& buffer, uint32_t& size){
+    void getOutput(std::unique_ptr<uint8_t[]>& buffer, uint32_t& size){
         size = 0;
         buffer = nullptr;
         if(outputArray_ && outputArraySize_ > 0){
             size = outputArraySize_;
-            buffer = std::make_unique<uint32_t[]>(outputArraySize_);
-            memcpy(buffer.get(), outputArray_.get(), outputArraySize_ * sizeof(uint32_t));
+            buffer = std::make_unique<uint8_t[]>(outputArraySize_);
+            memcpy(buffer.get(), outputArray_.get(), outputArraySize_ * sizeof(uint8_t));
         }
     }
 
@@ -49,8 +49,8 @@ public:
     void initInputBuffer(uint32_t rgb_size, uint32_t augmented_size){
         inputArraySize_ = rgb_size;
         if(inputArraySize_ > 0){
-            inputArray_ = std::make_unique<uint32_t[]>(inputArraySize_);
-            memset(inputArray_.get(), 0, inputArraySize_ * sizeof(uint32_t));
+            inputArray_ = std::make_unique<uint8_t[]>(inputArraySize_);
+            memset(inputArray_.get(), 0, inputArraySize_ * sizeof(uint8_t));
         }
 
         inputAugmentedArraySize_ = augmented_size;
@@ -63,15 +63,15 @@ public:
     void initOutputBuffer(uint32_t size){
         outputArraySize_ = size;
         if(outputArraySize_ > 0){
-            outputArray_ = std::make_unique<uint32_t[]>(outputArraySize_);
-            memset(outputArray_.get(), 0, size * sizeof(uint32_t));
+            outputArray_ = std::make_unique<uint8_t[]>(outputArraySize_);
+            memset(outputArray_.get(), 0, size * sizeof(uint8_t));
         }
     }
 
-    void copyInputArray(uint32_t *outArray, uint8_t* outAugmentedBuffer)
+    void copyInputArray(uint8_t *outArray, uint8_t* outAugmentedBuffer)
     {
         if(outArray && inputArray_ && inputArraySize_ > 0){
-            memcpy(outArray, inputArray_.get(), inputArraySize_ * sizeof(uint32_t));
+            memcpy(outArray, inputArray_.get(), inputArraySize_ * sizeof(uint8_t));
         } 
         
         if(outAugmentedBuffer && inputAugmentedArray_ && inputAugmentedArraySize_ > 0){
@@ -79,12 +79,12 @@ public:
         }
     }
 
-    void setInputBufferData(uint32_t *bufferData, uint32_t size){
+    void setInputBufferData(uint8_t *bufferData, uint32_t size){
         if(inputArray_ && inputArraySize_ > 0){
-            memset(inputArray_.get(), 0, inputArraySize_ * sizeof(uint32_t));
+            memset(inputArray_.get(), 0, inputArraySize_ * sizeof(uint8_t));
             if(size > 0){
                 uint32_t final_size = size > inputArraySize_ ? inputArraySize_ : size;
-                memcpy(inputArray_.get(), bufferData, final_size * sizeof(uint32_t));
+                memcpy(inputArray_.get(), bufferData, final_size * sizeof(uint8_t));
             }
         }
     }
@@ -99,11 +99,11 @@ public:
         }
     }
 
-    void setOutputBufferData(uint32_t *bufferData)
+    void setOutputBufferData(uint8_t *bufferData)
     {
         if(bufferData && outputArray_ && outputArraySize_ > 0){
-            memset(outputArray_.get(), 0, outputArraySize_ * sizeof(uint32_t));
-            memcpy(outputArray_.get(), bufferData, outputArraySize_ * sizeof(uint32_t));
+            memset(outputArray_.get(), 0, outputArraySize_ * sizeof(uint8_t));
+            memcpy(outputArray_.get(), bufferData, outputArraySize_ * sizeof(uint8_t));
         }
     }
 
@@ -156,9 +156,9 @@ public:
     }
     
 private:
-    std::unique_ptr<uint32_t[]> inputArray_;
+    std::unique_ptr<uint8_t[]> inputArray_;
     std::unique_ptr<uint8_t[]> inputAugmentedArray_;
-    std::unique_ptr<uint32_t[]> outputArray_;
+    std::unique_ptr<uint8_t[]> outputArray_;
     
     uint32_t inputArraySize_;
     uint32_t inputAugmentedArraySize_;
@@ -186,7 +186,7 @@ extern "C"{
         unityBridge::getBridge()->initOutputBuffer(size);
     }
 
-    void __stdcall getInputBufferCS(uint32_t* outBuffer, uint8_t* outAugmentedBuffer){
+    void __stdcall getInputBufferCS(uint8_t* outBuffer, uint8_t* outAugmentedBuffer){
         unityBridge::getBridge()->copyInputArray(outBuffer, outAugmentedBuffer);
         unityBridge::getBridge()->setNewBufferDataAvailable(false);
     }
@@ -215,7 +215,7 @@ extern "C"{
         unityBridge::getBridge()->setOutputHeight(height);
     }
 
-    void __stdcall setOutputBufferDataCS(uint32_t* bufferData){
+    void __stdcall setOutputBufferDataCS(uint8_t* bufferData){
         unityBridge::getBridge()->setOutputBufferData(bufferData);
     }
 
@@ -227,7 +227,7 @@ extern "C"{
 
 @implementation FrameworkLibAPI
 
-+ (void) getOutputBufferCpp:(std::unique_ptr<uint32_t[]>&)buffer size:(uint32_t&)size
++ (void) getOutputBufferCpp:(std::unique_ptr<uint8_t[]>&)buffer size:(uint32_t&)size
 {
     buffer = nullptr;
     size = 0;
@@ -238,7 +238,7 @@ extern "C"{
     return bridge->getOutput(buffer, size);
 }
 
-+ (void) setInputBufferCpp:(uint32_t*)buffer rgbSize:(uint32_t)rgb_size augmentedBuffer:(uint8_t*)augmentedBuffer augmentedSize:(uint32_t)augmented_size rotation:(int)rotation
++ (void) setInputBufferCpp:(uint8_t*)buffer rgbSize:(uint32_t)rgb_size augmentedBuffer:(uint8_t*)augmentedBuffer augmentedSize:(uint32_t)augmented_size rotation:(int)rotation
 {
     auto bridge = unityBridge::getBridge();
